feat(suse): replace CVRF feed with advisory CSAF updater
Switch SUSE security data from the legacy CVRF XML archive to the
upstream CSAF feed, storing trimmed native JSON under csaf/suse/.
