+++ .github/workflows/update.yml
@@ -116,15 +116,15 @@ jobs:
           commit-message: "CWE"
 
       - if: always()
-        name: SUSE CVRF
+        name: SUSE CSAF
         uses: ./.github/actions/update-source
         with:
-          target: suse-cvrf
+          target: suse-csaf
           client-id: ${{ secrets.SA_GH_VULN_LIST_UPDATE_GH_APP_CLIENT_ID }}
           private-key: ${{ secrets.SA_GH_VULN_LIST_UPDATE_GH_APP_PRIVATE_KEY }}
           owner: ${{ github.repository_owner }}
           repository: ${{ env.VULN_LIST_DIR }}
-          commit-message: "SUSE CVRF"
+          commit-message: "SUSE CSAF"
 
       - if: always()
         name: GitLab Advisory Database
