+++ main.go
@@ -31,7 +31,7 @@ import (
 	"github.com/aquasecurity/vuln-list-update/rocky"
 	"github.com/aquasecurity/vuln-list-update/rootio"
 	"github.com/aquasecurity/vuln-list-update/seal"
-	susecvrf "github.com/aquasecurity/vuln-list-update/suse/cvrf"
+	susecsaf "github.com/aquasecurity/vuln-list-update/suse/csaf"
 	"github.com/aquasecurity/vuln-list-update/ubuntu"
 	"github.com/aquasecurity/vuln-list-update/utils"
 	"github.com/aquasecurity/vuln-list-update/wolfi"
@@ -39,7 +39,7 @@ import (
 
 var (
 	target = flag.String("target", "", "update target (nvd, alpine, alpine-unfixed, redhat, redhat-oval, "+
-		"redhat-csaf-vex, debian, ubuntu, amazon, oracle-oval, suse-cvrf, photon, arch-linux, glad, cwe, osvdev, mariner, kevc, wolfi, "+
+		"redhat-csaf-vex, debian, ubuntu, amazon, oracle-oval, suse-csaf, photon, arch-linux, glad, cwe, osvdev, mariner, kevc, wolfi, "+
 		"chainguard, azure, openeuler, echo, minimos, eoldates, rootio)")
 	vulnListDir  = flag.String("vuln-list-dir", "", "vuln-list dir")
 	targetUri    = flag.String("target-uri", "", "alternative repository URI (only glad)")
@@ -108,10 +108,10 @@ func run() error {
 		if err := oc.Update(); err != nil {
 			return xerrors.Errorf("Oracle OVAL update error: %w", err)
 		}
-	case "suse-cvrf":
-		sc := susecvrf.NewConfig()
+	case "suse-csaf":
+		sc := susecsaf.NewConfig()
 		if err := sc.Update(); err != nil {
-			return xerrors.Errorf("SUSE CVRF update error: %w", err)
+			return xerrors.Errorf("SUSE CSAF update error: %w", err)
 		}
 	case "photon":
 		pc := photon.NewConfig()
