+++ suse/csaf/testdata/broken-csaf/suse-su-2019_0048-2.json
@@ -1,0 +1,1 @@
+{
