--- suse/cvrf/testdata/golden/openSUSE-SU-2015-0798-1.json
@@ -1,174 +1,0 @@
-{
-  "Title": "Security update for python-Pillow",
-  "Tracking": {
-    "ID": "openSUSE-SU-2015:0798-1",
-    "Status": "Final",
-    "Version": "1",
-    "InitialReleaseDate": "2015-04-22T11:03:36Z",
-    "CurrentReleaseDate": "2015-04-22T11:03:36Z",
-    "RevisionHistory": [
-      {
-        "Number": "1",
-        "Date": "2015-04-22T11:03:36Z",
-        "Description": "current"
-      }
-    ]
-  },
-  "Notes": [
-    {
-      "Text": "Security update for python-Pillow",
-      "Title": "Topic",
-      "Type": "Summary"
-    },
-    {
-      "Text": "python-pillow was updated to 2.7.0 to fix security issues and bugs.\n\nThe following vulnerabilities were fixed:\n\n* CVE-2014-9601: Remote attackers could cause a denial of service via a compressed text chunk in a PNG image that has a large size when it is decompressed.\n* CVE-2014-3598: Remote attackers could cause a denial of service using specially crafted image files via Jpeg2KImagePlugin\n* CVE-2014-3589: Remote attackers could cause a denial of service using specially crafted image files via IcnsImagePlugin\n",
-      "Title": "Details",
-      "Type": "General"
-    },
-    {
-      "Text": "The CVRF data is provided by SUSE under the Creative Commons License 4.0 with Attribution for Non-Commercial usage (CC-BY-NC-4.0).",
-      "Title": "Terms of Use",
-      "Type": "Legal Disclaimer"
-    }
-  ],
-  "ProductTree": {
-    "Relationships": [
-      {
-        "ProductReference": "python-Pillow-2.8.1-3.3.1",
-        "RelatesToProductReference": "openSUSE 13.2",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "python-Pillow-debuginfo-2.8.1-3.3.1",
-        "RelatesToProductReference": "openSUSE 13.2",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "python-Pillow-debugsource-2.8.1-3.3.1",
-        "RelatesToProductReference": "openSUSE 13.2",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "python-Pillow-tk-2.8.1-3.3.1",
-        "RelatesToProductReference": "openSUSE 13.2",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "python-Pillow-tk-debuginfo-2.8.1-3.3.1",
-        "RelatesToProductReference": "openSUSE 13.2",
-        "RelationType": "Default Component Of"
-      }
-    ]
-  },
-  "References": [
-    {
-      "URL": "http://lists.opensuse.org/opensuse-updates/2015-04/msg00056.html",
-      "Description": "E-Mail link for openSUSE-SU-2015:0798-1"
-    },
-    {
-      "URL": "https://www.suse.com/support/security/rating/",
-      "Description": "SUSE Security Ratings"
-    }
-  ],
-  "Vulnerabilities": [
-    {
-      "CVE": "CVE-2014-3589",
-      "Description": "PIL/IcnsImagePlugin.py in Python Imaging Library (PIL) and Pillow before 2.3.2 and 2.5.x before 2.5.2 allows remote attackers to cause a denial of service via a crafted block size.",
-      "Threats": [
-        {
-          "Type": "Impact",
-          "Severity": "moderate"
-        }
-      ],
-      "References": [
-        {
-          "URL": "https://www.suse.com/security/cve/CVE-2014-3589.html",
-          "Description": "CVE-2014-3589"
-        },
-        {
-          "URL": "https://bugzilla.suse.com/921566",
-          "Description": "SUSE Bug 921566"
-        }
-      ],
-      "ProductStatuses": [
-        {
-          "Type": "Fixed",
-          "ProductID": [
-            "openSUSE 13.2:python-Pillow-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-debuginfo-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-debugsource-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-tk-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-tk-debuginfo-2.8.1-3.3.1"
-          ]
-        }
-      ],
-      "CVSSScoreSets": {}
-    },
-    {
-      "CVE": "CVE-2014-3598",
-      "Description": "The Jpeg2KImagePlugin plugin in Pillow before 2.5.3 allows remote attackers to cause a denial of service via a crafted image.",
-      "Threats": [
-        {
-          "Type": "Impact",
-          "Severity": "moderate"
-        }
-      ],
-      "References": [
-        {
-          "URL": "https://www.suse.com/security/cve/CVE-2014-3598.html",
-          "Description": "CVE-2014-3598"
-        },
-        {
-          "URL": "https://bugzilla.suse.com/921566",
-          "Description": "SUSE Bug 921566"
-        }
-      ],
-      "ProductStatuses": [
-        {
-          "Type": "Fixed",
-          "ProductID": [
-            "openSUSE 13.2:python-Pillow-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-debuginfo-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-debugsource-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-tk-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-tk-debuginfo-2.8.1-3.3.1"
-          ]
-        }
-      ],
-      "CVSSScoreSets": {}
-    },
-    {
-      "CVE": "CVE-2014-9601",
-      "Description": "Pillow before 2.7.0 allows remote attackers to cause a denial of service via a compressed text chunk in a PNG image that has a large size when it is decompressed.",
-      "Threats": [
-        {
-          "Type": "Impact",
-          "Severity": "moderate"
-        }
-      ],
-      "References": [
-        {
-          "URL": "https://www.suse.com/security/cve/CVE-2014-9601.html",
-          "Description": "CVE-2014-9601"
-        },
-        {
-          "URL": "https://bugzilla.suse.com/921566",
-          "Description": "SUSE Bug 921566"
-        }
-      ],
-      "ProductStatuses": [
-        {
-          "Type": "Fixed",
-          "ProductID": [
-            "openSUSE 13.2:python-Pillow-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-debuginfo-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-debugsource-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-tk-2.8.1-3.3.1",
-            "openSUSE 13.2:python-Pillow-tk-debuginfo-2.8.1-3.3.1"
-          ]
-        }
-      ],
-      "CVSSScoreSets": {}
-    }
-  ]
-}
