--- suse/cvrf/testdata/golden/openSUSE-SU-2018-1633-1.json
@@ -1,200 +1,0 @@
-{
-  "Title": "Security update for glibc",
-  "Tracking": {
-    "ID": "openSUSE-SU-2018:1633-1",
-    "Status": "Final",
-    "Version": "1",
-    "InitialReleaseDate": "2018-06-09T08:29:23Z",
-    "CurrentReleaseDate": "2018-06-09T08:29:23Z",
-    "RevisionHistory": [
-      {
-        "Number": "1",
-        "Date": "2018-06-09T08:29:23Z",
-        "Description": "current"
-      }
-    ]
-  },
-  "Notes": [
-    {
-      "Text": "Security update for glibc",
-      "Title": "Topic",
-      "Type": "Summary"
-    },
-    {
-      "Text": "This update for glibc fixes the following issues:\n\nThis security issue was fixed:\n\n- Fixed an buffer overwrite issue in memcpy for Knights Landing CPUs\n  (boo#1092877, CVE-2018-11237)\n",
-      "Title": "Details",
-      "Type": "General"
-    },
-    {
-      "Text": "The CVRF data is provided by SUSE under the Creative Commons License 4.0 with Attribution for Non-Commercial usage (CC-BY-NC-4.0).",
-      "Title": "Terms of Use",
-      "Type": "Legal Disclaimer"
-    }
-  ],
-  "ProductTree": {
-    "Relationships": [
-      {
-        "ProductReference": "glibc-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-32bit-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-devel-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-devel-32bit-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-devel-static-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-devel-static-32bit-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-extra-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-html-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-i18ndata-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-info-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-locale-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-locale-32bit-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-profile-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-profile-32bit-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-testsuite-src-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-utils-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-utils-32bit-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "glibc-utils-src-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      },
-      {
-        "ProductReference": "nscd-2.26-lp150.11.3.2",
-        "RelatesToProductReference": "openSUSE Leap 15.0",
-        "RelationType": "Default Component Of"
-      }
-    ]
-  },
-  "References": [
-    {
-      "URL": "http://lists.opensuse.org/opensuse-security-announce/2018-06/msg00022.html",
-      "Description": "E-Mail link for openSUSE-SU-2018:1633-1"
-    },
-    {
-      "URL": "https://www.suse.com/support/security/rating/",
-      "Description": "SUSE Security Ratings"
-    }
-  ],
-  "Vulnerabilities": [
-    {
-      "CVE": "CVE-2018-11237",
-      "Description": "An AVX-512-optimized implementation of the mempcpy function in the GNU C Library (aka glibc or libc6) 2.27 and earlier may write data beyond the target buffer, leading to a buffer overflow in __mempcpy_avx512_no_vzeroupper.",
-      "Threats": [
-        {
-          "Type": "Impact",
-          "Severity": "moderate"
-        }
-      ],
-      "References": [
-        {
-          "URL": "https://www.suse.com/security/cve/CVE-2018-11237.html",
-          "Description": "CVE-2018-11237"
-        },
-        {
-          "URL": "https://bugzilla.suse.com/1092877",
-          "Description": "SUSE Bug 1092877"
-        },
-        {
-          "URL": "https://bugzilla.suse.com/1094154",
-          "Description": "SUSE Bug 1094154"
-        },
-        {
-          "URL": "https://bugzilla.suse.com/1118435",
-          "Description": "SUSE Bug 1118435"
-        }
-      ],
-      "ProductStatuses": [
-        {
-          "Type": "Fixed",
-          "ProductID": [
-            "openSUSE Leap 15.0:glibc-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-32bit-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-devel-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-devel-32bit-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-devel-static-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-devel-static-32bit-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-extra-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-html-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-i18ndata-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-info-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-locale-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-locale-32bit-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-profile-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-profile-32bit-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-testsuite-src-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-utils-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-utils-32bit-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:glibc-utils-src-2.26-lp150.11.3.2",
-            "openSUSE Leap 15.0:nscd-2.26-lp150.11.3.2"
-          ]
-        }
-      ],
-      "CVSSScoreSets": {}
-    }
-  ]
-}
